% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{fixAbundance}
\alias{fixAbundance}
\title{Fix total relative abundance to a set value}
\usage{
fixAbundance(plot, remain)
}
\arguments{
\item{plot}{a vector of species abundances}

\item{remain}{the total relative abundance value that \code{plot} should be
reduced to}
}
\value{
a list of two vectors of species abundances, the first with the
  original data provided by \code{plot}, the second where the total relative
  abundance is that set by \code{remain}
}
\description{
Takes a single plot (sample) and omits species, starting with the least
abundant, until a user-set total relative abundance in \code{plot} is reached
}
\keyword{internal}

