% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{functDiv}
\alias{functDiv}
\title{wrapper for FD::dbFD}
\usage{
functDiv(com, trait)
}
\arguments{
\item{com}{a community data matrix with species as columnes and plots (samples) as rows}

\item{trait}{vector, matrix, or data frame of trait values}
}
\value{
a list of different indices of functional diversity. See FD::dbFD for
   details.
}
\description{
subsets the community and trait data to contain the same species, as needed by dbFD
}
\keyword{internal}

