% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohrazeni}
\alias{ohrazeni}
\title{Community and trait data of wet meadows in Ohrazeni, Czech Republic}
\format{\code{ohrazeni$community} contains biomass data for 60 species in 12
  plots (samples). \code{ohrazeni$traits} contains values for 12 traits of
  the species in the community.}
\source{
Sampled in the field by Jan Leps and colleagues
}
\usage{
ohrazeni
}
\description{
A named list with two data frames, one containing community and the other
containing trait data.
}
\keyword{datasets}

