% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_reduceFD.R
\name{plot_reduceFD}
\alias{plot_reduceFD}
\title{Plot decay of rank correlations.}
\usage{
plot_reduceFD(reduced, index = c("FRich", "FEve", "FDis", "Rao", "CWM"))
}
\arguments{
\item{reduced}{object created by \code{\link{reduceFD}}.}

\item{index}{one of the indices calculated by \code{\link{reduceFD}}.}
}
\value{
A plot of decaying rank correlation with one line for each of the two
  reduction Scenarios. Only one plot for one functional diversity index is
  produced at a time.
}
\description{
Plots the decay of rank correlations between functional diversity indices of
original and reduced community data, as calculated by the
\code{\link{reduceFD}} function.
}

