% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{reducer}
\alias{reducer}
\title{reduce abundance of species in community matrix}
\usage{
reducer(com, remain = 0.5)
}
\arguments{
\item{com}{a community data matrix with species as columnes and plots (samples) as rows.}

\item{remain}{the total relative abundance value that com should be reduced to.}
}
\value{
a list of community data matrices: (1) Original community matrix, (2) Community matrix reduced according to Scenario 1, (3) Community matrix reduced according to Scenario 2
}
\description{
reduces the abundance of species in a community matrix according to a sample and a pool wise scenario for reduction
}
\keyword{internal}

