% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{sampler}
\alias{sampler}
\title{define species to be sampled}
\usage{
sampler(plot, avail_sp, thresh)
}
\arguments{
\item{plot}{a vector of species abundances}

\item{thresh}{a threshold value defining the minimum amount of relative
abundance that trait data should be available for}

\item{avail.sp}{a binary vector of 0 and 1 with the same number and order of
species as in \code{com} where 0 means that no trait data is available for
that species}
}
\value{
A character vector with the names of species that need to be sampled
}
\description{
defines what species in a plot (sample) should be sampled for measurement,
given what species have available trait data and how much of the relative
plot abundance should be covered
}
\keyword{internal}

