% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampleSpecies.R
\name{sampleSpecies}
\alias{sampleSpecies}
\title{Evaluate trait data availability for community data.}
\usage{
sampleSpecies(com, avail_sp, thresh = 0.8, sequential = TRUE)
}
\arguments{
\item{com}{a matrix containing the community data, with species as columns
and plots (samples) as rows.}

\item{avail_sp}{a binary vector of 0 and 1, either named with the same
species as in \code{com} or with the same number and order of species as in
\code{com}. 0 indicates that no trait data is available for a species.
Alternatively, a categorical or numeric vector of trait values but
containing NAs, where NA indicates species with missing trait data.}

\item{thresh}{a threshold value defining the minimum amount of relative
abundance that trait data should be available for.}

\item{sequential}{if \code{TRUE}, species to be sampled are first determined
by Scenario 1 and any species to be sampled in addition are determined by
Scenario 2. When \code{FALSE}, Scenario 2 gives the species to be sampled
in each plot to reach the threshold in each plot. See also details.}
}
\value{
a list with two or more elements, depending on the input data and
  arguments. See details.
}
\description{
This function evaluates for what proportion of species trait data is
available in a set of community data. Based on this information it also
provides the user with a list of species that should be sampled in order to
reach a user-defined threshold of data completeness, based on the abundance
data (e.g. percentage cover, individual counts, biomass) of the species.
}
\details{
There are different ways in which the data set is evaluated,
  according to the given data and the value of the \code{sequential}
  argument. In general, two different scenarios are applied to the data. In
  Scenario 1 (pool-wise scenario), the calculations are carried out on the
  basis of the pooled abundances of the species across the plots. In Scenario
  2 (plot-wise scenario), the calculations are carried out on each of the
  plots in the community data. Assessing the proportion of species that has
  trait data available with Scenario 1 therefor returns on single value,
  representing the total proportion of species' abundance for which trait
  data is available. With Scenario 2 this value is calculated for each of the
  plots.

  When \code{sequential} is set to \code{TRUE} the function first determines
  which species need to be sampled to reach the threshold according to
  Scenario 1, i.e. which species need to be sampled to ensure that the
  \strong{overall} (pooled) abundance covered is at least the value set by
  \code{thresh}. However, since this leads to the individual plots not
  necessarily reaching this threshold, subsequently species are determined by
  Scenario 2 (i.e. plot-wise) that need to be sampled \strong{in addition}
  to the ones determined by Scenario 1, in order to reach the threshold for
  each of the plots. Such a sampling strategy would be appropriate if one
  needs community trait values, i.e. it does not matter which individual plot
  the trait measurements come from. It can also occur that the threshold is
  already reached for each plot with Scenario 1, in which case a warning is
  shown and the value for Scenario 2 in the output is NULL.

  In the second case, \code{sequential} would be set to \code{FALSE} and then
  species to be sampled would be determined independent for Scenario 1 and
  Scenario 2. For Scenario 1, species determined would be the same, but for
  Scenario 2 species would be determined without taking into account species
  already suggested by Scenario 1. Hence, the function provides a full set of
  species that need to be sampled in each individual plot of the community.
  This would be an appropriate sampling strategy when for instance one would
  want to have plot-wise trait measurements, i.e. one could than take into
  account trait variation on the between-plot level, or even intraspecific
  level when several individual per species are measured per plot.

  All these examples discussed so far assume that species in the community
  data are represented by some sort of (semi)quantitative abundance measure
  (e.g. cover values, individual counts, biomass). Note that Scenario 2 can
  not be calculated when the community data is presence-absence only.

  \code{sampleSpecies} returns a list with three elements: \describe{
  \item{\code{$Original}}{Parameters of the original community data}
  \item{\code{$Scenario1}}{Parameters of the community data after applying
  Scenario 1} \item{\code{$Scenario2}}{Parameters of the community data after
  applying Scenario 2} } Each of these objects is itself a list, containing
  four objects, except for \code{$Original}, which only contains the latter
  three objects: \describe{\item{\code{$sample_species}}{The species to be
  sampled to reach the threshold. When \code{sequential} is \code{TRUE}, this
  gives a single vector of species to be sampled for both Scenario 1 and 2,
  with the vector under Scenario 2 given the species to be sampled in
  addition to these under Scenario 1. When \code{sequential} is \code{FALSE},
  this gives a list of vectors of species to be sampled, where each vector
  contains the species to be sampled in a given plot. \code{NULL} implies
  that the abundance threshold is being reached without sampling any
  (further) species than the ones that already have trait data available.}
  \item{\code{$com_available}}{The species available in the community data.
  If "NA", this species is not available.} \item{\code{$available_pool}}{The
  total proportional abundance available on the overall (pooled) community
  level.} \item{\code{$available_plot}}{The total proportional abundance
  available in each plot.} }
}
\examples{
# example community data and vector indicating trait availability for species
com <- matrix(c(50, 40, 30, 20, 0, 30, 0, 10, 0, 5, 10, 0, 0, 0, 5), 3, 5)
colnames(com) <- as.character(c(1:5))
avail <- c(1, 0, 1, 0, 1)

# run example with different argument settings
sampleSpecies(com, avail, 0.95, sequential = TRUE)
sampleSpecies(com, avail, 0.8, sequential = FALSE)

# run with the ohrazeni wet meadow data set
data(ohrazeni)
sampleSpecies(ohrazeni$vegetationdata, rbinom(61, 1, 0.5), 0.95, sequential = FALSE)
}
\references{
Majekova, M., et al. Evaluating functional diversity: missing
  trait data and the importance of species abundance structure and data
  transformation. Unpublished.
}

