% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohrazeni}
\alias{ohrazeni}
\title{Vegetation and trait data of wet meadows in Ohrazeni, Czech Republic}
\format{\code{ohrazeni$community} contains biomass data for 60 species in 12
  plots (samples). \code{ohrazeni$traits} contains values for 12 traits of
  the species in the community.}
\source{
Sampled in the field by Jan Leps and colleagues.
}
\usage{
ohrazeni
}
\description{
A named list with two data frames. The first one containing vegetation data
in form of above-ground biomass with species as columns and plots as rows.
The second dataframe contains trait data for the species (species as rows,
traits as columns).
}
\references{
Lepš, J., de Bello, F., Šmilauer, P., and Doležal, J. (2011)
  Community trait response to environment: disentangling species turnover vs
  intraspecific trait variability effects. Ecography 35, 856–863.

  Lepš, J. (2014) Scale- and time-dependent effects of fertilization, mowing
  and dominant removal on a grassland community during a 15-year experiment.
  Journal of Applied Ecology 51, 978–987.
}
\keyword{datasets}

