% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reduceFD.R
\name{reduceFD}
\alias{reduceFD}
\title{Functional diversity indices for original and reduced community data.}
\usage{
reduceFD(com, traits, reduction = 0.05, remain = 0.5)
}
\arguments{
\item{com}{a matrix containing the community data, with species as columns
and plots (samples) as rows.}

\item{traits}{a vector, data frame or matrix containing trait values for the
species in \code{com}.}

\item{reduction}{the relative abundance removed in each reduction step.}

\item{remain}{the amount of total relative abundance retained in the
community. Gives an error if this value would lead to plots having less
than two species.}
}
\value{
A list with five elements. See details.
}
\description{
\code{reduceFD} uses \code{\link[FD]{dbFD}} to calculate functional diversity
indices for original and reduced community data for two different scenarios
(pool-wise and plot-wise). Also, rank correlations between the FD indices of
original and reduced data are calculated (Majekova et al., unpubl.).
}
\details{
The function returns a list of five elements, containing the
  following objects: \describe{\item{\code{$rem_seq}}{A vector containing the
  sequence of reduction steps used to calculate functional diversities. The
  last (smallest) value in this vector is equal to \code{remain}.}
  \item{\code{$FD_reduced_scenario1}}{This object contains a list of the same
  length as \code{$rem_seq}, where each element contains the functional
  diversity indices for each of the reduction
  steps.}\item{\code{$FD_reduced_scenario1}}{The same as the previous object,
  but for Scenario 2}\item{\code{$rank_cor_scenario1}}{The rank correlations
  between the functional diversity indices of each of the reduction steps
  (for Scenario 1) and the functional diversity indices of the original data.
  This value demonstrates, for each functional diversity index separately,
  how the index from increasingly reduced communities contain less and less
  "true" information about the actual functional diversities across the plots
  (samples).}\item{\code{$rank_cor_scenario2}}{The same as the previous
  object, but for Scenario 2.}}
}
\examples{
data(ohrazeni)
red <- reduceFD(ohrazeni$vegetationdata,
     ohrazeni$traitdata[, "height", drop = F],
     reduction = 0.1, remain = 0.7)
plot_reduceFD(red, index = "Rao")
}
\references{
Majekova, Maria; et al. Evaluating functional diversity: missing
  trait data and the importance of species abundance structure and data
  transformation. Unpublished.
}
\seealso{
\code{\link[FD]{dbFD}} for details on functional diversity
  calculation, \code{\link{plot.reduceFD}} for a graphical visualisation of
  the output of reduceFD.
}

